Creación de la ruta /verifyToken para validar el token y devolver el usuario en caso de éxito

+++ backend/src/models/reviews.js
@@ -8,7 +8,7 @@ const Reviews = sequelize.define('Reviews', {
     primaryKey: true,
     allowNull: false,
   },
-  rating: DataTypes.TINYINT(5),
+  rating: DataTypes.INTEGER,
   title: DataTypes.TEXT,
   comment: DataTypes.TEXT,
 })
